+++ extensions/healthie/actions/index.ts
@@ -18,6 +18,7 @@ import { updatePatient } from './updatePatient'
 import { getPatient } from './getPatient'
 import { applyTagToPatient } from './applyTagToPatient'
 import { updatePatientQuickNote } from './updatePatientQuickNote'
+import { createMetricEntry } from './createMetricEntry'
 
 export const actions = {
   createAppointment,
@@ -49,4 +50,5 @@ export const actions = {
   completeTask,
   assignPatientToGroup,
   updatePatientQuickNote,
+  createMetricEntry,
 }